+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * Updates untilNow parameter on postAddCompany [#74](https://github.com/xing/XNGAPIClient/pull/74)
 * Add GET for sent contact requests [#82](https://github.com/xing/XNGAPIClient/pull/82)
 * Keychain bug fixes [#86](https://github.com/xing/XNGAPIClient/pull/86)
+* Add Events calls [#84](https://github.com/xing/XNGAPIClient/pull/84)